@@ -2,10 +2,12 @@ import 'package:flutter/material.dart';
 import 'login_screen.dart';
 import 'registration_screen.dart';
 import 'package:animated_text_kit/animated_text_kit.dart';
-import 'package:flash_chat/components/rounded_button.dart';
+import 'package:asmvp/components/rounded_button.dart';
 
 class WelcomeScreen extends StatefulWidget {
   static const String id = 'welcome_screen';
+
+  const WelcomeScreen({super.key});
 
   @override
   _WelcomeScreenState createState() => _WelcomeScreenState();
@@ -13,8 +15,10 @@ class WelcomeScreen extends StatefulWidget {
 
 class _WelcomeScreenState extends State<WelcomeScreen>
     with SingleTickerProviderStateMixin {
-  AnimationController controller;
-  Animation animation;
+  bool isDarkMode = true;
+
+  AnimationController? controller;
+  Animation? animation;
 
   @override
   void initState() {
@@ -23,23 +27,23 @@ class _WelcomeScreenState extends State<WelcomeScreen>
     controller =
         AnimationController(duration: Duration(seconds: 1), vsync: this);
     animation = ColorTween(begin: Colors.blueGrey, end: Colors.white)
-        .animate(controller);
-    controller.forward();
-    controller.addListener(() {
+        .animate(controller!);
+    controller?.forward();
+    controller?.addListener(() {
       setState(() {});
     });
   }
 
   @override
   void dispose() {
-    controller.dispose();
+    controller?.dispose();
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: animation.value,
+      backgroundColor: animation?.value,
       body: Padding(
         padding: EdgeInsets.symmetric(horizontal: 24.0),
         child: Column(
@@ -50,16 +54,34 @@ class _WelcomeScreenState extends State<WelcomeScreen>
               children: <Widget>[
                 Hero(
                   tag: 'logo',
-                  child: Container(
-                    child: Image.asset('images/logo.png'),
+                  child: SizedBox(
                     height: 60.0,
+                    child: Image.asset('images/logo.png'),
                   ),
                 ),
-                TypewriterAnimatedTextKit(
-                  text: ['Flash Chat'],
-                  textStyle: TextStyle(
-                    fontSize: 45.0,
-                    fontWeight: FontWeight.w900,
+                Expanded(
+                  child: AnimatedTextKit(
+                    key: ValueKey<bool>(isDarkMode),
+                    animatedTexts: [
+                      TypewriterAnimatedText(
+                        isDarkMode ? 'Armaan Solutions' : 'Armaan Solutions',
+                        cursor: isDarkMode ? '>' : '<',
+                        textStyle: TextStyle(
+                            fontSize: 45,
+                            fontWeight: FontWeight.w900,
+                            color: isDarkMode
+                                ? Colors.amber[300]
+                                : Colors.grey[850]),
+                        speed: const Duration(milliseconds: 100),
+                      ),
+                      //      TypewriterAnimatedTextKit(
+                      //   text: ['Flash Chat'],
+                      //   textStyle: TextStyle(
+                      //     fontSize: 45.0,
+                      //     fontWeight: FontWeight.w900,
+                      //   ),
+                      // ),
+                    ],
                   ),
                 ),
               ],
