@@ -1,10 +1,20 @@
 import 'package:flutter/material.dart';
-import 'package:flash_chat/screens/welcome_screen.dart';
-import 'package:flash_chat/screens/login_screen.dart';
-import 'package:flash_chat/screens/registration_screen.dart';
-import 'package:flash_chat/screens/chat_screen.dart';
+import 'package:asmvp/screens/welcome_screen.dart';
+import 'package:asmvp/screens/login_screen.dart';
+import 'package:asmvp/screens/registration_screen.dart';
+import 'package:asmvp/screens/chat_screen.dart';
 
-void main() => runApp(FlashChat());
+import 'package:firebase_core/firebase_core.dart';
+import 'firebase_options.dart';
+
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+
+  await Firebase.initializeApp(
+    options: DefaultFirebaseOptions.currentPlatform,
+  );
+  runApp(FlashChat());
+}
 
 class FlashChat extends StatelessWidget {
   @override
